refactor(ast): remove ast terminology, simplify and clarify expressions
