+++ packages/fast-element/src/expression.ts
@@ -1,19 +1,29 @@
-export interface IExpressionSource {
-    getExpression(): IExpression;
-}
-
+/**
+ * Provides additional contextual information available to arrow functions
+ * evaluated in the context of a template update.
+ */
 export interface IEvaluationContext<T = any> {
     event: Event;
     parent: T;
     index: number;
 }
 
+/**
+ * A simple abstraction of an expression which can be evaluated as part of
+ * a template update.
+ */
 export interface IExpression {
     evaluate(scope: unknown, context?: IEvaluationContext): unknown;
 }
 
+/**
+ * The signature of an arrow function capable of being evluated as part of a template update.
+ */
 export type Getter<T = any, K = any> = (model: T, context: IEvaluationContext) => K;
 
+/**
+ * A basic implementation of IExpression, which wraps a Getter function.
+ */
 export class AccessScopeExpression<T = any, K = any> implements IExpression {
     constructor(public getter: Getter<T, K>) {}
 
@@ -30,6 +40,10 @@ export class AccessScopeExpression<T = any, K = any> implements IExpression {
     }
 }
 
+/**
+ * An implementation of IExpression which interpolates string literal values with
+ * Getter functions to produce a final string.
+ */
 export class InterpolationExpression implements IExpression {
     constructor(private parts: (string | Getter)[]) {}
